Cambio de implementación del Merssene twister
Se utiliza la implementación "mtwist" del algoritmo Merssene twister para la
generación de números aleatorios, de http://www.cs.hmc.edu/~geoff/mtwist.html

Como hace uso de `inline' hay que compilar con la opción -fgnu89-inline
poqrue la semántica de esa palabra en C99 y C89 es distinta.

La función al_azar ahora devuelve unsigned long int porque el índice de
un individuo puede superar al rango de unsigned int.

+++ funciones.h
@@ -13,6 +13,7 @@
 #include <string.h>
 #include <time.h>
 #include <unistd.h>
+#include "mtwist.h"
 
 #define UNUSED(x) (void)(x)
 
@@ -119,7 +120,7 @@ void cruza_ciclica (struct individuos_s *restrict madre,
 
 void mutacion (struct individuos_s *restrict individuo);
 
-unsigned int al_azar (const unsigned int min, const unsigned int max);
+unsigned long al_azar (const unsigned int min, const unsigned int max);
 
 double al_azar_d (const double min, const double max);
